requirements.txt
openai==0.28.1
python-dotenv==1.0.0
fastapi==0.108.0
uvicorn[standard]==0.25.0
tiktoken==0.5.2

sentence-transformers==2.2.2
qdrant-client==1.9.0